psx-keycloak-auth-middleware/pyproject.toml
[tool.poetry]
authors = ["kutlu69", "taeyma41", "dekeuk93"]
description = "Keycloak Auth Middleware"
name = "keycloak-auth-middleware-svc"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
python = "~3.12"
uvicorn = "^0.29.0"
python-keycloak = "^4.3.0"
sse-starlette = "^2.1.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.generators.dependencies]
datamodel-code-generator = "^0.25.6"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
